+++ lib/slining.rb
@@ -1,4 +1,5 @@
-require 'slining/version'
-require 'slining/generators/app_generator'
-require 'slining/actions'
-require 'slining/app_builder'
+require "slining/version"
+require "slining/generators/app_generator"
+require "slining/actions"
+require "slining/adapters/heroku"
+require "slining/app_builder"